[Feat] 유저 프로필 컨텐츠 담당 UserProfileCard 파일 대신 UserProfileContent 파일 생성 및 사용
